fix[FE]: Fix not matching queue
