+++ frontend/src/hooks/useGame.ts
@@ -1,5 +1,4 @@
 import { useContext, useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { SocketContext } from 'contexts/socket';
 
 type ReturnType = [
@@ -13,22 +12,12 @@ export default function useGame(): ReturnType {
   const [isWating, setIsWating] = useState(false);
   const [alertCode, setAlertCode] = useState<string | null>(null);
   const socket = useContext(SocketContext);
-  const navigate = useNavigate();
 
   useEffect(() => {
     if (!isWating) return;
 
     socket.on('ping', () => {
       socket.emit('pong');
-    });
-
-    socket.on('queue', (data: { text: string; gameId?: number }) => {
-      if (data.gameId) {
-        navigate(`/game/${data.gameId}/setting`);
-      } else {
-        setAlertCode(data.text);
-        setIsWating(false);
-      }
     });
 
     return () => {
